[portinglayer] renamed tcc_driver to tcc_mode, changed SetInnerLight to SetAlarm, updated matter_driver_gpio_level_irq_handler function, removed api demonstration at example_matter_refrigerator.cpp, and inititalize the tcc mode at matter_driver_refrigerator_set_startup_value function
