+++ component/common/application/matter/driver/refrigerator_driver.h
@@ -7,16 +7,16 @@
 class MatterRefrigerator
 {
 public:
-    void Init(PinName gpioLight);
+    void Init(PinName outputGpio);
     void deInit(void);
     uint16_t GetMode();
     void SetMode(uint16_t newMode);
     uint8_t GetDoorStatus(void);
     void SetDoorStatus(uint8_t temp);
-    void SetInnerLight(void);
+    void SetAlarm(void);
     
 private:
-    gpio_t innerLight;
+    gpio_t alarmGpio;
     uint16_t mode;
     uint8_t doorStatus;
 };
