+++ component/common/application/matter/driver/tcc_mode.cpp
@@ -16,7 +16,7 @@
  *    limitations under the License.
  */
 #include <app-common/zap-generated/attributes/Accessors.h>
-#include "tcc_driver.h"
+#include "tcc_mode.h"
 
 using namespace chip::app::Clusters;
 using namespace chip::app::Clusters::RefrigeratorAndTemperatureControlledCabinetMode;
